Factor out TranspilerRepository into its own module.
